feat: remove old about affine
