+++ apps/web/src/components/affine/setting-modal/general-setting/about/icons.tsx
